Move out synchronisation stuff from DataTypesJSON
