+++ haskell-lsp-types/src/Language/Haskell/LSP/Types/DataTypesJSON.hs
@@ -22,14 +22,11 @@ import           Data.Scientific                            (floatingOrInteger)
 import           Data.Text                                  (Text)
 import qualified Data.Text                                  as T
 import           Language.Haskell.LSP.Types.ClientCapabilities
-import           Language.Haskell.LSP.Types.Diagnostic
 import           Language.Haskell.LSP.Types.Common
-import           Language.Haskell.LSP.Types.Location
 import           Language.Haskell.LSP.Types.LspId
 import           Language.Haskell.LSP.Types.Method
 import           Language.Haskell.LSP.Types.Progress
 import           Language.Haskell.LSP.Types.ServerCapabilities
-import           Language.Haskell.LSP.Types.TextDocument
 import           Language.Haskell.LSP.Types.Uri
 import           Language.Haskell.LSP.Types.Utils
 import           Language.Haskell.LSP.Types.WorkspaceEdit
@@ -534,337 +531,6 @@ deriveJSON lspOptions ''ConfigurationParams
 
 -- ---------------------------------------------------------------------
 {-
-DidOpenTextDocument Notification
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#didopentextdocument-notification
-
-The document open notification is sent from the client to the server to signal
-newly opened text documents. The document's truth is now managed by the client
-and the server must not try to read the document's truth using the document's
-uri.
-
-Notification:
-
-    method: 'textDocument/didOpen'
-    params: DidOpenTextDocumentParams defined as follows:
-
-interface DidOpenTextDocumentParams {
-    /**
-     * The document that was opened.
-     */
-    textDocument: TextDocumentItem;
-}
-
-Registration Options: TextDocumentRegistrationOptions
--}
-
-data DidOpenTextDocumentParams =
-  DidOpenTextDocumentParams {
-    _textDocument :: TextDocumentItem
-  } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''DidOpenTextDocumentParams
-
--- ---------------------------------------------------------------------
-{-
-DidChangeTextDocument Notification
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#didchangetextdocument-notification
-
-    Changed: The document change notification is sent from the client to the
-    server to signal changes to a text document. In 2.0 the shape of the params
-    has changed to include proper version numbers and language ids.
-
-Notification:
-
-    method: 'textDocument/didChange'
-    params: DidChangeTextDocumentParams defined as follows:
-
-interface DidChangeTextDocumentParams {
-    /**
-     * The document that did change. The version number points
-     * to the version after all provided content changes have
-     * been applied.
-     */
-    textDocument: VersionedTextDocumentIdentifier;
-
-    /**
-     * The actual content changes.
-     */
-    contentChanges: TextDocumentContentChangeEvent[];
-}
-
-/**
- * An event describing a change to a text document. If range and rangeLength are omitted
- * the new text is considered to be the full content of the document.
- */
-interface TextDocumentContentChangeEvent {
-    /**
-     * The range of the document that changed.
-     */
-    range?: Range;
-
-    /**
-     * The length of the range that got replaced.
-     */
-    rangeLength?: number;
-
-    /**
-     * The new text of the document.
-     */
-    text: string;
-}
--}
-data TextDocumentContentChangeEvent =
-  TextDocumentContentChangeEvent
-    { _range       :: Maybe Range
-    , _rangeLength :: Maybe Int
-    , _text        :: Text
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions { omitNothingFields = True } ''TextDocumentContentChangeEvent
-
--- -------------------------------------
-
-data DidChangeTextDocumentParams =
-  DidChangeTextDocumentParams
-    { _textDocument   :: VersionedTextDocumentIdentifier
-    , _contentChanges :: List TextDocumentContentChangeEvent
-    } deriving (Show,Read,Eq)
-
-deriveJSON lspOptions ''DidChangeTextDocumentParams
-
-{-
-New in 3.0
-----------
-
-Registration Options: TextDocumentChangeRegistrationOptions defined as follows:
-
-/**
- * Descibe options to be used when registered for text document change events.
- */
-export interface TextDocumentChangeRegistrationOptions extends TextDocumentRegistrationOptions {
-        /**
-         * How documents are synced to the server. See TextDocumentSyncKind.Full
-         * and TextDocumentSyncKindIncremental.
-         */
-        syncKind: number;
-}
--}
-
-makeExtendingDatatype "TextDocumentChangeRegistrationOptions"
-  [''TextDocumentRegistrationOptions]
-  [("_syncKind", [t| TextDocumentSyncKind |])]
-
-deriveJSON lspOptions ''TextDocumentChangeRegistrationOptions
-
--- ---------------------------------------------------------------------
-{-
-
-New in 3.0
-----------
-
-WillSaveTextDocument Notification
-
-The document will save notification is sent from the client to the server before
-the document is actually saved.
-
-Notification:
-
-    method: 'textDocument/willSave'
-    params: WillSaveTextDocumentParams defined as follows:
-
-/**
- * The parameters send in a will save text document notification.
- */
-export interface WillSaveTextDocumentParams {
-        /**
-         * The document that will be saved.
-         */
-        textDocument: TextDocumentIdentifier;
-
-        /**
-         * The 'TextDocumentSaveReason'.
-         */
-        reason: number;
-}
-
-/**
- * Represents reasons why a text document is saved.
- */
-export namespace TextDocumentSaveReason {
-
-        /**
-         * Manually triggered, e.g. by the user pressing save, by starting debugging,
-         * or by an API call.
-         */
-        export const Manual = 1;
-
-        /**
-         * Automatic after a delay.
-         */
-        export const AfterDelay = 2;
-
-        /**
-         * When the editor lost focus.
-         */
-        export const FocusOut = 3;
-}
-Registration Options: TextDocumentRegistrationOptions
--}
-
-data TextDocumentSaveReason
-  = SaveManual
-         -- ^ Manually triggered, e.g. by the user pressing save, by starting
-         -- debugging, or by an API call.
-  | SaveAfterDelay -- ^ Automatic after a delay
-  | SaveFocusOut   -- ^ When the editor lost focus
-  deriving (Show, Read, Eq)
-
-instance A.ToJSON TextDocumentSaveReason where
-  toJSON SaveManual     = A.Number 1
-  toJSON SaveAfterDelay = A.Number 2
-  toJSON SaveFocusOut   = A.Number 3
-
-instance A.FromJSON TextDocumentSaveReason where
-  parseJSON (A.Number 1) = pure SaveManual
-  parseJSON (A.Number 2) = pure SaveAfterDelay
-  parseJSON (A.Number 3) = pure SaveFocusOut
-  parseJSON _            = mempty
-
-data WillSaveTextDocumentParams =
-  WillSaveTextDocumentParams
-    { _textDocument :: TextDocumentIdentifier
-    , _reason       :: TextDocumentSaveReason
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''WillSaveTextDocumentParams
-
--- ---------------------------------------------------------------------
-{-
-New in 3.0
-----------
-
-WillSaveWaitUntilTextDocument Request
-
-The document will save request is sent from the client to the server before the
-document is actually saved. The request can return an array of TextEdits which
-will be applied to the text document before it is saved. Please note that
-clients might drop results if computing the text edits took too long or if a
-server constantly fails on this request. This is done to keep the save fast and
-reliable.
-
-Request:
-
-    method: 'textDocument/willSaveWaitUntil'
-    params: WillSaveTextDocumentParams
-
-Response:
-
-    result: TextEdit[]
-    error: code and message set in case an exception happens during the willSaveWaitUntil request.
-
-Registration Options: TextDocumentRegistrationOptions
--}
-
--- ---------------------------------------------------------------------
-{-
-DidSaveTextDocument Notification
-
-The document save notification is sent from the client to the server when the document was saved in the client.
-
-Client Capability:
-
-property name (optional): textDocument.synchronization.didSave
-property type: boolean
-The capability indicates that the client supports textDocument/didSave notifications.
-
-Server Capability:
-
-property name (optional): textDocumentSync.save
-property type: boolean | SaveOptions where SaveOptions is defined as follows:
-export interface SaveOptions {
-	/**
-	 * The client is supposed to include the content on save.
-	 */
-	includeText?: boolean;
-}
-The capability indicates that the server is interested in textDocument/didSave notifications.
-
-Registration Options: TextDocumentSaveRegistrationOptions defined as follows:
-
-export interface TextDocumentSaveRegistrationOptions extends TextDocumentRegistrationOptions {
-	/**
-	 * The client is supposed to include the content on save.
-	 */
-	includeText?: boolean;
-}
-Notification:
-
-method: 'textDocument/didSave'
-params: DidSaveTextDocumentParams defined as follows:
-interface DidSaveTextDocumentParams {
-	/**
-	 * The document that was saved.
-	 */
-	textDocument: TextDocumentIdentifier;
-
-	/**
-	 * Optional the content when saved. Depends on the includeText value
-	 * when the save notification was requested.
-	 */
-	text?: string;
-}
--}
-data DidSaveTextDocumentParams =
-  DidSaveTextDocumentParams
-    { _textDocument :: TextDocumentIdentifier
-    , _text         :: Maybe Text
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''DidSaveTextDocumentParams
-
-makeExtendingDatatype "TextDocumentSaveRegistrationOptions"
-  [''TextDocumentRegistrationOptions]
-  [("_saveOptions", [t| SaveOptions |])]
-
-deriveJSON lspOptions ''TextDocumentSaveRegistrationOptions
-
--- ---------------------------------------------------------------------
-{-
-DidCloseTextDocument Notification
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#didclosetextdocument-notification
-
-The document close notification is sent from the client to the server when the
-document got closed in the client. The document's truth now exists where the
-document's uri points to (e.g. if the document's uri is a file uri the truth now
-exists on disk).
-
-    Changed: In 2.0 the params are of type DidCloseTextDocumentParams which
-    contains a reference to a text document.
-
-Notification:
-
-    method: 'textDocument/didClose'
-    params: DidCloseTextDocumentParams defined as follows:
-
-interface DidCloseTextDocumentParams {
-    /**
-     * The document that was closed.
-     */
-    textDocument: TextDocumentIdentifier;
-}
--}
-data DidCloseTextDocumentParams =
-  DidCloseTextDocumentParams
-    { _textDocument :: TextDocumentIdentifier
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''DidCloseTextDocumentParams
-
--- ---------------------------------------------------------------------
-{-
 DidChangeWatchedFiles Notification
 
 https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#didchangewatchedfiles-notification
@@ -978,14 +644,6 @@ interface PublishDiagnosticsParams {
 }
 -}
 
-data PublishDiagnosticsParams =
-  PublishDiagnosticsParams
-    { _uri         :: Uri
-    , _diagnostics :: List Diagnostic
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''PublishDiagnosticsParams
-
 
 -- ---------------------------------------------------------------------
 {-
@@ -1039,18 +697,6 @@ error: code and message set in case an exception happens during the definition r
 -}
 
 -- {"jsonrpc":"2.0","id":1,"method":"textDocument/definition","params":{"textDocument":{"uri":"file:///tmp/Foo.hs"},"position":{"line":1,"character":8}}}
-
-data LocationResponseParams = SingleLoc Location | MultiLoc [Location]
-  deriving (Eq,Read,Show)
-
-instance A.ToJSON LocationResponseParams where
-  toJSON (SingleLoc x) = toJSON x
-  toJSON (MultiLoc xs) = toJSON xs
-
-instance A.FromJSON LocationResponseParams where
-  parseJSON xs@(A.Array _) = MultiLoc <$> parseJSON xs
-  parseJSON x              = SingleLoc <$> parseJSON x
-
 
 -- ---------------------------------------------------------------------
 {-
